requirements.txt
streamlit>=1.28.0
aiohttp>=3.8.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
wikipedia>=1.4.0
duckduckgo-search>=4.1.0
arxiv>=2.1.0
feedparser>=6.0.10
python-dateutil>=2.8.0
pandas>=2.2.0
plotly>=5.18.0
numpy>=1.26.0
python-dotenv>=1.0.0
rich>=13.7.0
ollama>=0.1.6 